Reland "Fix dependency loop with flex"
I'm adding a prebuilt_build_tool module for flex so that it can be used
by genrules. But since flex uses flex during it's own build, and
prebuilt_build_tool can't handle the loop, explicitly change the name to
"flex_bin", and use stem so that the final binary is still called flex.

Bug: 128690776
Test: treehugger
Change-Id: I6a818cbd297f951552d2890374787eb40c074463