Fix dependency loop with flex
I'm adding a prebuilt_build_tool module for flex so that it can be used
by genrules. But since flex uses flex during it's own build, and
prebuilt_build_tool can't handle the loop, explicitly change the name to
"flex_bin", and use stem so that the final binary is still called flex.

Bug: 128690776
Test: treehugger
Change-Id: Ibcb257e7b39dd23b203605ae163980ca854e95b9